+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Please add your entries according to this format.
 
 ## Unreleased
 - Print the paths of files ktfmt failed to analyze instead of only how many files failed to be analyzed (#469) 
+- Make Gradle tasks `KtfmtBaseTask`, `KtfmtCheckTask` and `KtfmtFormatTask` reusable from other Gradle plugins (#467)
 
 ## Version 0.24.0 _(2025-09-08)_
 - Add support for android projects with the `com.android.kotlin.multiplatform.library` plugin (#422)